make faster
Former-commit-id: 5d16119383452dfb7a219150e1891a6ade4a07d0

+++ install.sh
@@ -13,9 +13,11 @@ else
 	sudo yum install -y docker
 	sudo usermod -a -G docker ec2-user
 fi
+	
+if [[ ! -f /usr/bin/docker-compose ]]; then
+	curl -L https://github.com/docker/compose/releases/download/1.7.0/docker-compose-`uname -s`-`uname -m` > ./docker-compose
+	sudo mv ./docker-compose /usr/bin/docker-compose
+	sudo chmod +x /usr/bin/docker-compose
+fi
 
 sudo service docker start
-
-curl -L https://github.com/docker/compose/releases/download/1.7.0/docker-compose-`uname -s`-`uname -m` > ./docker-compose
-sudo mv ./docker-compose /usr/bin/docker-compose
-sudo chmod +x /usr/bin/docker-compose
